Added VNH simulator

+++ TDD/drivers/gpio.h
@@ -15,9 +15,9 @@
 extern "C" {
 #endif
 
-void gpio_set(uint16_t gpio_nb_);
+void gpio_set(uint32_t gpio_nb_);
 
-uint8_t gpio_get(uint16_t gpio_nb_);
+uint8_t gpio_get(uint32_t gpio_nb_);
 
 void gpio_toggle(uint32_t gpio_nb_);
